requirements.txt
fastapi[standard]==0.115.12
pydantic_settings==2.7.1
sqlmodel==0.0.22
psycopg==3.2.4
uvicorn[standard]==0.34.2